+++ docs/plan/CHANGELOG.md
@@ -1,5 +1,9 @@
 # 追加したドキュメント
 
+2026-08-18
+
+- docs/plan/earth-geography-heightmaps.md
+
 2026-08-16
 
 - docs/simulation/frontier-start-modes.md